refactoring folders and structure
