+++ app/api/auth_service/auth.py
@@ -7,7 +7,7 @@ from jose import jwt, JWTError
 from sqlalchemy.orm import Session
 from sqlalchemy import select
 from sqlalchemy.exc import NoResultFound
-from app.api.models.models import User
+from app.core.models import User
 from app.api.routes.users.schemas import UserViewDTO
